backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]
gunicorn
python-multipart
sqlalchemy==2.0.45
asyncpg
aiosqlite
passlib[bcrypt]
bcrypt
PyJWT
python-dotenv==1.0.1
pydantic
aiohttp
qrcode[pil]
Pillow==11.1.0